Added NG, GB and UZBEK EditorCategories
