+++ addons/garmsar_boys/CfgEditorCategories.hpp
@@ -1,0 +1,9 @@
+class CfgEditorCategories {
+
+
+    class Spagistan_NG_Category // Category class, you point to it in editorCategory property
+        {
+            displayName = "Spagistan National Guard"; // Name visible in the list
+            faction = "SPAG_NG";
+        };
+};
