+++ addons/garmsar_boys/config.cpp
@@ -3,3 +3,4 @@
 #include "CfgFactionClasses.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgGroups.hpp"
+#include "CfgEditorCategories.hpp"
