+++ addons/national_guard/CfgEditorCategories.hpp
@@ -1,0 +1,9 @@
+class CfgEditorCategories {
+
+
+	class Spagistan_GC_Category // Category class, you point to it in editorCategory property
+	{
+		displayName = "Spagistan Garmsar Cadets"; // Name visible in the list
+		faction = "SPAG_GC";
+	};
+};
