+++ addons/uzbek/CfgEditorCategories.hpp
@@ -1,0 +1,8 @@
+class CfgEditorCategories {
+
+	class O_UZAF_Category // Category class, you point to it in editorCategory property
+	{
+		displayName = "Uzbek Armed Forces"; // Name visible in the list
+		faction = "O_UZAF";
+	};
+};
